Use dict in place of ordereddict in marshal
This aim to fix our problems of memory, it's a complicated but let say
that each ordereddict create at least one cyclic list: a list than
contain itself. The garbage collector of python handle that kind of
thing pretty well, but for performance reason collection of long lived
objects is not done too often. As explained here:
https://github.com/python/cpython/blob/master/Modules/gcmodule.c#L86
collection of old objects is only done when at least 25% of the old
objects has not lived a full collection yet.

On long request, typically isochrone, the ordereddict from the response
live long enough to be considered has old objects, for most fast
requests objects don't live long enough to become "old", so we almost
never trigger a full collection exempt on isochrone request.
So memory don't grow without control, but a lot of the memory allocated
for an isochrone request isn't really released until the next isochrone
request.

Test have been made with an isochrone request on fr-idf with a max
duration of 1H30:

|          | dict  | orderedDict |
| memory   | 180mo | 500mo       |
| duration | 12sec | 17sec       |